stm32f7/nucleo-144: refactor bringup logic
