+++ boards/arm/stm32f7/nucleo-144/src/stm32_boot.c
@@ -83,13 +83,8 @@ void stm32_boardinitialize(void)
 #ifdef CONFIG_BOARD_LATE_INITIALIZE
 void board_late_initialize(void)
 {
-#if defined(CONFIG_NSH_LIBRARY) && !defined(CONFIG_BOARDCTL)
-  /* Perform NSH initialization here instead of from the NSH.  This
-   * alternative NSH initialization is necessary when NSH is ran in
-   * user-space but the initialization function must run in kernel space.
-   */
+  /* Perform board-specific initialization */
 
-  board_app_initialize(0);
-#endif
+  stm32_bringup();
 }
 #endif
